+++ src/RequestType.js
@@ -183,6 +183,8 @@ export default class RequestType {
                 return "GetAccountDetails";
             case RequestType.EthereumTransaction:
                 return "EthereumTransaction";
+            case RequestType.NodeStakeUpdate:
+                return "NodeStakeUpdate";
             default:
                 return `UNKNOWN (${this._code})`;
         }
@@ -337,6 +339,8 @@ export default class RequestType {
                 return RequestType.GetAccountDetails;
             case 84:
                 return RequestType.EthereumTransaction;
+            case 85:
+                return RequestType.NodeStakeUpdate;
         }
 
         throw new Error(
@@ -704,3 +708,8 @@ RequestType.GetAccountDetails = new RequestType(83);
  * Ethereum Transaction
  */
 RequestType.EthereumTransaction = new RequestType(84);
+
+/**
+ * Updates the staking info at the end of staking period to indicate new staking period has started.
+ */
+RequestType.NodeStakeUpdate = new RequestType(84);