+++ src/contract/ContractCreateTransaction.js
@@ -62,6 +62,10 @@ export default class ContractCreateTransaction extends Transaction {
      * @param {Uint8Array} [props.constructorParameters]
      * @param {string} [props.contractMemo]
      * @param {number} [props.maxAutomaticTokenAssociations]
+     * @param {AccountId | string} [props.stakedAccountId]
+     * @param {Long | number} [props.stakedNodeId]
+     * @param {boolean} [props.declineStakingReward]
+     * @param {AccountId} [props.autoRenewAccountId]
      */
     constructor(props = {}) {
         super();
@@ -128,6 +132,29 @@ export default class ContractCreateTransaction extends Transaction {
 
         this._defaultMaxTransactionFee = new Hbar(20);
 
+        /**
+         * @private
+         * @type {?AccountId}
+         */
+        this._stakedAccountId = null;
+
+        /**
+         * @private
+         * @type {?Long}
+         */
+        this._stakedNodeId = null;
+
+        /**
+         * @private
+         * @type {boolean}
+         */
+        this._declineStakingReward = false;
+
+        /**
+         * @type {?AccountId}
+         */
+        this._autoRenewAccountId = null;
+
         if (props.bytecodeFileId != null) {
             this.setBytecodeFileId(props.bytecodeFileId);
         }
@@ -149,6 +176,7 @@ export default class ContractCreateTransaction extends Transaction {
         }
 
         if (props.proxyAccountId != null) {
+            // eslint-disable-next-line deprecation/deprecation
             this.setProxyAccountId(props.proxyAccountId);
         }
 
@@ -168,6 +196,22 @@ export default class ContractCreateTransaction extends Transaction {
             this.setMaxAutomaticTokenAssociations(
                 props.maxAutomaticTokenAssociations
             );
+        }
+
+        if (props.stakedAccountId != null) {
+            this.setStakedAccountId(props.stakedAccountId);
+        }
+
+        if (props.stakedNodeId != null) {
+            this.setStakedNodeId(props.stakedNodeId);
+        }
+
+        if (props.declineStakingReward != null) {
+            this.setDeclineStakingReward(props.declineStakingReward);
+        }
+
+        if (props.autoRenewAccountId != null) {
+            this.setAutoRenewAccountId(props.autoRenewAccountId);
         }
     }
 
@@ -234,6 +278,19 @@ export default class ContractCreateTransaction extends Transaction {
                 maxAutomaticTokenAssociations:
                     create.maxAutomaticTokenAssociations != null
                         ? create.maxAutomaticTokenAssociations
+                        : undefined,
+                stakedAccountId:
+                    create.stakedAccountId != null
+                        ? AccountId._fromProtobuf(create.stakedAccountId)
+                        : undefined,
+                stakedNodeId:
+                    create.stakedNodeId != null
+                        ? create.stakedNodeId
+                        : undefined,
+                declineStakingReward: create.declineReward == true,
+                autoRenewAccountId:
+                    create.autoRenewAccountId != null
+                        ? AccountId._fromProtobuf(create.autoRenewAccountId)
                         : undefined,
             }),
             transactions,
@@ -345,6 +402,7 @@ export default class ContractCreateTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @returns {?AccountId}
      */
     get proxyAccountId() {
@@ -352,6 +410,7 @@ export default class ContractCreateTransaction extends Transaction {
     }
 
     /**
+     * @deprecated
      * @param {AccountId | string} proxyAccountId
      * @returns {this}
      */
@@ -373,6 +432,10 @@ export default class ContractCreateTransaction extends Transaction {
     }
 
     /**
+     * An account to charge for auto-renewal of this contract. If not set, or set to an
+     * account with zero hbar balance, the contract's own hbar balance will be used to
+     * cover auto-renewal fees.
+     *
      * @param {Duration | Long | number} autoRenewPeriod
      * @returns {this}
      */
@@ -443,6 +506,84 @@ export default class ContractCreateTransaction extends Transaction {
     }
 
     /**
+     * @returns {?AccountId}
+     */
+    get stakedAccountId() {
+        return this._stakedAccountId;
+    }
+
+    /**
+     * @param {AccountId | string} stakedAccountId
+     * @returns {this}
+     */
+    setStakedAccountId(stakedAccountId) {
+        this._requireNotFrozen();
+        this._stakedAccountId =
+            typeof stakedAccountId === "string"
+                ? AccountId.fromString(stakedAccountId)
+                : stakedAccountId;
+
+        return this;
+    }
+
+    /**
+     * @returns {?Long}
+     */
+    get stakedNodeId() {
+        return this._stakedNodeId;
+    }
+
+    /**
+     * @param {Long | number} stakedNodeId
+     * @returns {this}
+     */
+    setStakedNodeId(stakedNodeId) {
+        this._requireNotFrozen();
+        this._stakedNodeId = Long.fromValue(stakedNodeId);
+
+        return this;
+    }
+
+    /**
+     * @returns {boolean}
+     */
+    get declineStakingRewards() {
+        return this._declineStakingReward;
+    }
+
+    /**
+     * @param {boolean} declineStakingReward
+     * @returns {this}
+     */
+    setDeclineStakingReward(declineStakingReward) {
+        this._requireNotFrozen();
+        this._declineStakingReward = declineStakingReward;
+
+        return this;
+    }
+
+    /**
+     * @returns {?AccountId}
+     */
+    get autoRenewAccountId() {
+        return this._autoRenewAccountId;
+    }
+
+    /**
+     * @param {string | AccountId} autoRenewAccountId
+     * @returns {this}
+     */
+    setAutoRenewAccountId(autoRenewAccountId) {
+        this._requireNotFrozen();
+        this._autoRenewAccountId =
+            typeof autoRenewAccountId === "string"
+                ? AccountId.fromString(autoRenewAccountId)
+                : autoRenewAccountId;
+
+        return this;
+    }
+
+    /**
      * @param {Client} client
      */
     _validateChecksums(client) {
@@ -502,6 +643,16 @@ export default class ContractCreateTransaction extends Transaction {
             constructorParameters: this._constructorParameters,
             memo: this._contractMemo,
             maxAutomaticTokenAssociations: this._maxAutomaticTokenAssociations,
+            stakedAccountId:
+                this.stakedAccountId != null
+                    ? this.stakedAccountId._toProtobuf()
+                    : null,
+            stakedNodeId: this.stakedNodeId,
+            declineReward: this.declineStakingRewards,
+            autoRenewAccountId:
+                this._autoRenewAccountId != null
+                    ? this._autoRenewAccountId._toProtobuf()
+                    : null,
         };
     }