+++ src/exports.js
@@ -18,6 +18,7 @@
  * ‍
  */
 
+export { default as Cache } from "./Cache.js";
 export {
     BadKeyError,
     BadMnemonicError,