+++ src/sync.js
@@ -2,18 +2,39 @@ import axios from "axios";
 import Cache from "./Cache.js";
 import Logger from "js-logger";
 
-const sync = (async () => {
-    // http://time.google.com:80 doesn't actually give us an NTP response, instead it returns
-    // a 302 redirected response. However, it does contain a `date` header which we can use.
+/**
+ * @returns {Promise<void>}
+ */
+export async function syncFunction() {
+    // Make a "fake" account balance query and extract the date header from the response
     try {
-        const response = await axios.get("http://time.google.com", {
-            maxRedirects: 0,
+        const response = await axios.post(
+            // "https://myhbarwallet.com:443",
+            "https://grpc-web.myhbarwallet.com:443/proto.CryptoService/cryptoGetBalance",
+            {},
+            {
+                maxRedirects: 0,
+                headers: {
+                    "content-type": "application/grpc-web+proto",
+                    "x-user-agent": "hedera-sdk-js/v2",
+                    "x-grpc-web": "1",
+                },
 
-            // By default a status of 302 is considered an erring status
-            validateStatus: function (status) {
-                return status < 500;
-            },
-        });
+                // By default a status of 302 is considered an erring status
+                validateStatus: function (status) {
+                    return status < 500;
+                },
+            }
+        );
+
+        if (
+            response == null ||
+            response.headers == null ||
+            response.headers.date == null ||
+            Number.isNaN(response.headers.date)
+        ) {
+            return;
+        }
 
         const currentTime = Math.round(Date.now() / 1000);
 
@@ -32,6 +53,10 @@ const sync = (async () => {
             }`
         );
     }
+}
+
+const sync = (async () => {
+    await syncFunction();
 })();
 
 export default sync;