+++ test/unit/sync.js
@@ -1,0 +1,7 @@
+import { syncFunction } from "../../src/sync.js";
+
+describe("sync", function () {
+    it("is able to fetch HEAD", async function () {
+        await syncFunction();
+    });
+});